Created JumpingOnTheCloudsRevisited.java
Aerith is playing a cloud hopping game. In this game, there are sequentially numbered clouds that can be thunderheads or cumulus clouds. Her character must jump from cloud to cloud until it reaches the start again.

To play, Aerith is given an array of clouds,c  and an energy level 100. She starts from c[0] and uses 1 unit of energy to make a jump of size k to cloud . If Aerith lands on a thundercloud, , her energy () decreases by 2  additional units. The game ends when Aerith lands back on cloud .

Given the values of , , and the configuration of the clouds as an array , can you determine the final value of  after the game ends?